Rename: add postRegister folder
